refactor: Only use pooled dbContexts
